+++ src/main/java/blackjack/dto/CurrentCardsDTO.java
@@ -1,8 +1,8 @@
 package blackjack.dto;
 
-import blackjack.domain.Card;
-import blackjack.domain.Dealer;
-import blackjack.domain.Player;
+import blackjack.domain.card.Card;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Player;
 
 import java.util.List;
