+++ src/test/java/blackjack/domain/DeckTest.java
@@ -1,6 +1,8 @@
 package blackjack.domain;
 
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
